requirements.txt
Flask
pytesseract
openai>=1.3.5
python-dotenv
Pillow
